Moved model loading to shared library
