+++ lessons/11. First Rendering System/Cargo.toml
@@ -11,3 +11,4 @@ vulkano-win = "0.29.0"
 vulkano-shaders = "0.29.0"
 bytemuck = { version = "1.7", features = ["derive", "extern_crate_std", "min_const_generics"] }
 nalgebra-glm = "0.16.0"
+model_loader = { path = "../../model_loader" }
