--- lessons/11. First Rendering System/src/model.rs
@@ -1,104 +1,0 @@
-// Copyright (c) 2022 taidaesal
-// Licensed under the MIT license
-// <LICENSE-MIT or http://opensource.org/licenses/MIT>
-
-#![allow(dead_code)]
-use super::obj_loader::{Loader, NormalVertex};
-
-use nalgebra_glm::{identity, inverse_transpose, rotate_normalized_axis, translate, TMat4, TVec3};
-
-/// Holds our data for a renderable model, including the model matrix data
-///
-/// Note: When building an instance of `Model` the loader will assume that
-/// the input obj file is in clockwise winding order. If it is already in
-/// counter-clockwise winding order, call `.invert_winding_order(false)`
-/// when building the `Model`.
-pub struct Model {
-    data: Vec<NormalVertex>,
-    translation: TMat4<f32>,
-    rotation: TMat4<f32>,
-    model: TMat4<f32>,
-    normals: TMat4<f32>,
-    // we might call multiple translation/rotation calls
-    // in between asking for the model matrix. This lets
-    // only recreate the model matrix when needed.
-    requires_update: bool,
-}
-
-pub struct ModelBuilder {
-    file_name: String,
-    custom_color: [f32; 3],
-    invert: bool,
-}
-
-impl ModelBuilder {
-    fn new(file: String) -> ModelBuilder {
-        ModelBuilder {
-            file_name: file,
-            custom_color: [1.0, 0.35, 0.137],
-            invert: true,
-        }
-    }
-
-    pub fn build(self) -> Model {
-        let loader = Loader::new(self.file_name.as_str(), self.custom_color, self.invert);
-        Model {
-            data: loader.as_normal_vertices(),
-            translation: identity(),
-            rotation: identity(),
-            model: identity(),
-            normals: identity(),
-            requires_update: false,
-        }
-    }
-
-    pub fn color(mut self, new_color: [f32; 3]) -> ModelBuilder {
-        self.custom_color = new_color;
-        self
-    }
-
-    pub fn file(mut self, file: String) -> ModelBuilder {
-        self.file_name = file;
-        self
-    }
-
-    pub fn invert_winding_order(mut self, invert: bool) -> ModelBuilder {
-        self.invert = invert;
-        self
-    }
-}
-
-impl Model {
-    pub fn new(file_name: &str) -> ModelBuilder {
-        ModelBuilder::new(file_name.into())
-    }
-
-    pub fn data(&self) -> Vec<NormalVertex> {
-        self.data.clone()
-    }
-
-    pub fn model_matrices(&mut self) -> (TMat4<f32>, TMat4<f32>) {
-        if self.requires_update {
-            self.model = self.translation * self.rotation;
-            self.normals = inverse_transpose(self.model);
-            self.requires_update = false;
-        }
-        (self.model, self.normals)
-    }
-
-    pub fn rotate(&mut self, radians: f32, v: TVec3<f32>) {
-        self.rotation = rotate_normalized_axis(&self.rotation, radians, &v);
-        self.requires_update = true;
-    }
-
-    pub fn translate(&mut self, v: TVec3<f32>) {
-        self.translation = translate(&self.translation, &v);
-        self.requires_update = true;
-    }
-
-    /// Return the model's rotation to 0
-    pub fn zero_rotation(&mut self) {
-        self.rotation = identity();
-        self.requires_update = true;
-    }
-}
