+++ model_loader/src/loader/vertex.rs
